add sipgate pro routing
If a user has a Sipgate Pro account and registers all numbers as SIP gateways, the RAND() function gives any calling number whatever is set in qualified.
The new ORDER BY changes this behavior and serves first the one wich should be used through the gates. To get it working the name of the gateways per MSN has to be the MSN (qualified number).